[mdns] change `PublishService()` to get encoded TXT data as input
This commit updates `Mdns::Publisher::PublishService()` to take
an encoded TXT data as its input. This harmonizes the `Publish`
API to be similar to the `Subscribe()` which also uses encoded
TXT data and make it easier to use with advertising proxy and
TREL (avoid extra step of decoding already encoded TXT data
so to call the `PublishService()`).